move checks_mock.ex to test/support
