+++ config/e2e_test.exs
@@ -6,7 +6,8 @@ config :plausible, PlausibleWeb.Endpoint,
 
 config :plausible,
   paddle_api: Plausible.Billing.DevPaddleApiMock,
-  google_api: Plausible.Google.API.Mock
+  google_api: Plausible.Google.API.Mock,
+  verification_checks_mod: Plausible.InstallationSupport.Verification.ChecksMock
 
 config :phoenix, :stacktrace_depth, 20
 config :phoenix, :plug_init_mode, :runtime
