+++ extra/lib/plausible/installation_support/installation_support.ex
@@ -13,13 +13,7 @@ defmodule Plausible.InstallationSupport do
       "Plausible Verification Agent - if abused, contact support@plausible.io"
     end
 
-    def verification_checks_mod do
-      if Mix.env() in [:dev, :e2e_test] do
-        Plausible.InstallationSupport.Verification.ChecksMock
-      else
-        Plausible.InstallationSupport.Verification.Checks
-      end
-    end
+    def verification_checks_mod, do: Application.fetch_env!(:plausible, :verification_checks_mod)
   else
     def user_agent() do
       "Plausible Community Edition"
